rgb.h: switch to uint8_t

+++ rgb.h
@@ -4,10 +4,9 @@
 
 namespace Cygni {
     struct RGB {
-        // TODO: uint8_t
-        int red;
-        int green;
-        int blue;
+        uint8_t red;
+        uint8_t green;
+        uint8_t blue;
 
         bool operator==(const RGB &other) {
             return red == other.red && green == other.green && blue == other.blue;
